+++ index.html
@@ -213,36 +213,36 @@
       </div>
     </div>
 
-    <div id="member-review-container">
-      <span>참가자 후기</span>
-      <div class="review-card-list">
-        <ds-review-item
+    <ds-reviews-section>
+      <ds-hero>참가자 후기</ds-hero>
+      <ds-review-list>
+        <ds-review
           authorImgSrc="images/human_color.png"
           content="해외취업을 목표로 리트코드 스터디 모임에 참여한 것은 좋은 선택이었습니다. 특히 혼자 공부하는 것보다 멤버들과 함께 문제를 풀며 시너지를 낼 수 있었습니다. 모임 중에 다양한 질문을 주고받으며 실력을 쌓을 수 있었고, 해외 취업 준비에 실질적인 도움을 받을 수 있었습니다."
           author="John Doe"
-        ></ds-review-item>
-        <ds-review-item
+        ></ds-review>
+        <ds-review
           authorImgSrc="images/human_color.png"
           content="해외취업을 목표로 리트코드 스터디 모임에 참여한 것은 좋은 선택이었습니다. 특히 혼자 공부하는 것보다 멤버들과 함께 문제를 풀며 시너지를 낼 수 있었습니다. 모임 중에 다양한 질문을 주고받으며 실력을 쌓을 수 있었고, 해외 취업 준비에 실질적인 도움을 받을 수 있었습니다."
           author="John Doe"
-        ></ds-review-item>
-        <ds-review-item
+        ></ds-review>
+        <ds-review
           authorImgSrc="images/human_color.png"
           content="해외취업을 목표로 리트코드 스터디 모임에 참여한 것은 좋은 선택이었습니다. 특히 혼자 공부하는 것보다 멤버들과 함께 문제를 풀며 시너지를 낼 수 있었습니다. 모임 중에 다양한 질문을 주고받으며 실력을 쌓을 수 있었고, 해외 취업 준비에 실질적인 도움을 받을 수 있었습니다."
           author="John Doe"
-        ></ds-review-item>
-        <ds-review-item
+        ></ds-review>
+        <ds-review
           authorImgSrc="images/human_color.png"
           content="해외취업을 목표로 리트코드 스터디 모임에 참여한 것은 좋은 선택이었습니다. 특히 혼자 공부하는 것보다 멤버들과 함께 문제를 풀며 시너지를 낼 수 있었습니다. 모임 중에 다양한 질문을 주고받으며 실력을 쌓을 수 있었고, 해외 취업 준비에 실질적인 도움을 받을 수 있었습니다."
           author="John Doe"
-        ></ds-review-item>
-        <ds-review-item
+        ></ds-review>
+        <ds-review
           authorImgSrc="images/human_color.png"
           content="해외취업을 목표로 리트코드 스터디 모임에 참여한 것은 좋은 선택이었습니다. 특히 혼자 공부하는 것보다 멤버들과 함께 문제를 풀며 시너지를 낼 수 있었습니다. 모임 중에 다양한 질문을 주고받으며 실력을 쌓을 수 있었고, 해외 취업 준비에 실질적인 도움을 받을 수 있었습니다."
           author="John Doe"
-        ></ds-review-item>
-      </div>
-    </div>
+        ></ds-review>
+      </ds-review-list>
+    </ds-reviews-section>
 
     <ds-steps-section>
       <ds-hero>스터디 참여방법</ds-hero>
