+++ main.js
@@ -6,7 +6,8 @@ import "./components/hero.js";
 import "./components/icon-link-list.js";
 import "./components/image.js";
 import "./components/intro-section.js";
-import "./components/review-item.js";
+import "./components/review.js";
+import "./components/review-section.js";
 import "./components/seo-meta-tag/seo-meta-tag.js";
 import "./components/step.js";
 import "./components/step-section.js";
